+++ src/mathed/MathSupport.h
@@ -45,14 +45,14 @@ int mathed_char_width(FontInfo const &, char_type c);
 int mathed_char_kerning(FontInfo const &, char_type c);
 
 // thickness added to decorations because of the zoom value
-int mathed_deco_thickness(MetricsBase & mb);
+double mathed_deco_thickness(MetricsBase & mb);
 
 // add the thickness of the decorations. num_w and num_h are a multiplier
 // in case there are several decorations, horizontally or vertically
 void mathed_deco_metrics(MetricsBase & mb, Dimension & dim,
                          int num_w = 1, int num_h = 1);
 
-void mathed_draw_deco(PainterInfo & pi, int x, int y, int w, int h,
+void mathed_draw_deco(PainterInfo & pi, double x, double y, double w, double h,
                       docstring const & name);
 
 void mathed_string_dim(FontInfo const & font,
